# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RunConfigIn"
          },
          "context_docs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContextDocIn"
            },
            "maxItems": 5,
            "title": "Context Docs",
            "type": "array"
          },
          "from_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Run"
          },
          "harness": {
            "default": "claude",
            "enum": [
              "claude",
              "codex",
              "demo"
            ],
            "title": "Harness",
            "type": "string"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "RunList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first API for Oracle run orchestration and artifact discovery.",
    "title": "Oracle GUI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "A page of runs, newest first.\n\nDemo runs are hidden unless asked for, by `include_demo` or by naming the harness:\nthey are practice runs and would otherwise crowd out the scientist's real work.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "harness",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Harness"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "show_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Show Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_demo",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Demo",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "enum": [
                "recent",
                "oldest",
                "title",
                "calls"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "default": false,
              "title": "Mine",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Create a run and start its supervisor. Returns as soon as the process is alive.\n\nAlways a **new** run, `from_run` included: that field copies an earlier run's question,\nprompt, config and context documents, and the run it creates starts from no hypotheses,\nno ratings and no guidance. Extending a run that already exists is\n`POST /runs/{id}/controls {action: \"continue\"}`, which keeps the whole idea pool.\n\nThe 409 is the harness lane: one active run per harness, enforced by a partial unique\nindex rather than a check the API could race against. Its details name the run holding\nthe lane only when that run is visible to this user; otherwise they report capacity\nwithout exposing another workspace.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "delete": {
        "description": "Soft delete: the run disappears from every list and 404s directly; the row stays.",
        "operationId": "delete_run_api_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Run",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      },
      "patch": {
        "description": "Rename or archive a run. Neither field touches its lifecycle, budget or content.",
        "operationId": "patch_run_api_runs__run_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/context": {
      "get": {
        "operationId": "get_run_context_api_runs__run_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContextDocRef"
                  },
                  "title": "Response Get Run Context Api Runs  Run Id  Context Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Context",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/controls": {
      "post": {
        "description": "Pause, resume, continue, stop, finish or force-stop a run (plan C4's action table).\n\nCooperative actions are acknowledged, not completed: the lifecycle in the response is\nthe one the run is in *now*, and it moves when the supervisor acts on the flag at its\nnext step boundary.\n\nTwo of them start a process instead of setting a flag, and both answer with the run's\nnew lifecycle rather than its old one. `resume` picks a paused run back up; `continue`\ngives a run that already ended `add_rounds` more rounds, in place and keeping everything\nit learned — optionally raising its call ceiling, and optionally raising or removing the\ncost ceiling that may have been what stopped it. Both take the harness lane again, so\nboth can 409 exactly as a launch does.",
        "operationId": "control_run_api_runs__run_id__controls_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Control Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/detail": {
      "get": {
        "description": "Everything a run workspace needs in one round trip.",
        "operationId": "get_run_detail_api_runs__run_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Detail",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/export.md": {
      "get": {
        "description": "The run as a downloadable markdown report — the Report tab's Download button.",
        "operationId": "export_run_api_runs__run_id__export_md_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/graph": {
      "get": {
        "description": "The run's idea genealogy — nodes, descent edges and the legend's facts, in one trip.\n\nRead-only and assembled from rows the run already has: what the Ideas tab draws is the\nattrition story, and fetching it as four calls would be four chances to render half a\nrun.",
        "operationId": "get_run_graph_api_runs__run_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Graph",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/hypotheses": {
      "get": {
        "operationId": "list_hypotheses_api_runs__run_id__hypotheses_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "elo",
              "enum": [
                "elo",
                "hid",
                "recent"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HypothesisRow"
                  },
                  "title": "Response List Hypotheses Api Runs  Run Id  Hypotheses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Hypotheses",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}